+++ src/plugin/packer/manifest/__tests__/sourcelist.test.ts
